pyproject.toml
[tool.poetry]
name = "syahello"
version = "0.1.0"
description = "Simple Message"
authors = ["Kevin Cheng <isolatedp920@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.1.0"
python-dotenv = "^1.0.1"
flask-wtf = "^1.2.2"
flask-sqlalchemy = "^3.1.1"
pytz = "^2024.2"
pyyaml = "^6.0.2"
pymysql = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
